Move struct `syntactic_closure` into header `syntactic_closure.hpp`
